pet-clinic-data/src/main/java/guru/springframework/sfgpetclinic/model/Pet.java
package guru.springframework.sfgpetclinic.model;

import lombok.*;
import org.springframework.format.annotation.DateTimeFormat;

import javax.persistence.*;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;
import java.time.LocalDate;
import java.util.HashSet;
import java.util.Set;

@Setter
@Getter
@NoArgsConstructor
@AllArgsConstructor
@Entity
@EqualsAndHashCode(exclude = "owner", callSuper = false)
@Table(name = "pets")
public class Pet extends BaseEntity {

    @NotBlank
    private String name;

    @ManyToOne()
    @JoinColumn(name = "pet_type_id")
    @NotNull
    private PetType petType;
    @ManyToOne
    @JoinColumn(name = "owner_id")
    private Owner owner;
    @Column(name = "birth_date")
    @NotNull
    @DateTimeFormat(pattern = "yyyy-MM-dd")
    private LocalDate birthDate;

    @OneToMany(cascade = CascadeType.ALL, mappedBy = "pet")
    private Set<Visit> visits = new HashSet<>();

    public void addVisit(Visit visit) {
        if (visit.isNew()) {
            visits.add(visit);
        }
        visit.setPet(this);
    }
}
